Python Proxy
What these code does it , it acts as a proxy or bridge between our salesopt.ai website and activepieces website running in our cloud server. It takes post request with user name,email and password and use it to login the user automatically to activepieces container. It will receive token from activepieces which then it stores it in localstorage to allow user login automatically without need to login to two different website. 1. In our website and 2. In activepieces website. Once user login then it take request from user and then convert it to correct url and then send it to activepieces. when activepieces container sends response in reuturn then it well translate the reponse and send it to the user. So it acts like a bridge.