Solved the Signin and logout issue
I was trying to use token from global token and was redundant information. Add logout  code which added removal of cokkies which contains tokens as well as platform id and token id and also taking care of other local storage and session storage also along with index databases. Token is excess everywhere using cokkies instead of using global variable. I also remove some redundant function which was updating the token but was actually using token from previous session and throwing error. As the token is update , I also account for different project id as well platform id used in api as well as query parameters in API , if they don't match then we change the poject id as well as platform id to the recent id obtained from the cokkies which was set when user sign in or sign up.